Add charBar with localstorage
