+++ src/js/components/barChart.js
@@ -7,23 +7,27 @@ class TransactionProcessor {
 
   // Extraer el mes y año
   static getMonthYearFromDate(dateString) {
-    const [day, month, year] = dateString.split('/');
-    return `${year}-${String(month).padStart(2, '0')}`;
+    const date = new Date(dateString);
+
+    return `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(
+      2,
+      '0'
+    )}`;
   }
 
   // Filtrar por tipo, mes y año para obtener el monto mensual
   filterByTypeAndMonth(type, selectedYear) {
     return this.transactions.reduce((accumulator, transaction) => {
       const monthYear = TransactionProcessor.getMonthYearFromDate(
-        transaction['Fecha']
+        transaction['date']
       );
-      const amount = transaction['Monto'];
 
-      if (!monthYear.startsWith(selectedYear)) return accumulator; // Filtrar solo el año deseado
+      const amount = +transaction['amount'];
+
+      if (!monthYear.startsWith(selectedYear)) return accumulator;
 
       if (!accumulator[monthYear]) accumulator[monthYear] = 0;
-      accumulator[monthYear] +=
-        transaction['Tipo de Registro'] === type ? amount : 0;
+      accumulator[monthYear] += transaction['type'] === type ? amount : 0;
 
       return accumulator;
     }, {});
@@ -43,7 +47,7 @@ class TransactionProcessor {
 }
 
 // Clase para manejar el gráfico de barras
-class BarChart {
+export class BarChart {
   constructor(canvasId) {
     this.canvasId = canvasId;
     this.chartInstance = null;
@@ -53,7 +57,7 @@ class BarChart {
   renderChart(labels, datasets) {
     const ctx = document.getElementById(this.canvasId)?.getContext('2d');
 
-    if (this.chartInstance) this.chartInstance.destroy(); // Evitar gráficos duplicados
+    if (this.chartInstance) this.chartInstance.destroy();
 
     this.chartInstance = new Chart(ctx, {
       type: 'bar',
@@ -86,31 +90,3 @@ class BarChart {
     }
   }
 }
-
-import transactions from '../../assets/data/transactions.json';
-
-// Inicializar el gráfico cuando la página carga
-document.addEventListener('DOMContentLoaded', () => {
-  const currentYear = new Date().getFullYear(); // Año actual dinámico
-  const labels = Array.from({ length: 12 }, (_, i) =>
-    String(i + 1).padStart(2, '0')
-  );
-
-  const datasetConfig = [
-    {
-      label: 'Gastos',
-      type: 'Gasto',
-      color: 'rgba(213, 213, 247, 1)',
-      borderRadius: 3,
-    },
-    {
-      label: 'Ingresos',
-      type: 'Ingreso',
-      color: 'rgba(47, 44, 216, 1)',
-      borderRadius: 3,
-    },
-  ];
-
-  const myChart = new BarChart('cash-flow-chart');
-  myChart.renderData(transactions, labels, datasetConfig, currentYear);
-});
